requirements.txt
openai>=1.0.0
elasticsearch>=8.0.0,<9.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0

